requirements.txt
torch
torchvision
numpy
opencv-python
matplotlib
tqdm
Pillow
